코드 컨벤션에 맞게 이쁘게 리팩토링
